APM-3433 Move auth code helper functions
